requirements.txt
aiohttp==3.11.11
anyio==4.8.0
anthropic
azure-ai-inference==1.0.0b8
azure-ai-projects==1.0.0b5
azure-ai-ml
azure-core==1.32.0
azure-identity
colorama==0.4.6
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
dotenv
executing==2.2.0
httpcore==1.0.7
httpx==0.28.1
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jwt
matplotlib-inline==0.1.7
msal==1.31.1
numpy==2.2.2
openai==1.60.0
openapi-core==0.19.4
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
pybars4==0.9.13
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMeta3==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==308
rich
referencing==0.35.1
requests==2.32.3
semantic-kernel==1.19.0
streamlit
typing
urllib3==2.3.0
uvicorn
ldap3==2.9.1
fastapi== 0.101.1
msgraph-sdk
fastmcp==2.13.0.2
asyncio
ftfy==6.1.1